ENH: break MatchResults into smaller components
